Itemvalues now update if searched. Graph of searched item will be shown
Every kind of Impling Jar can now be looked up by typing its name in code.

The code will then look if the name is found in the database, and if it is, give the corresponding itemID.

If an itemID is found for the item, the API will return price and time values for that item, the 14 latest values of that item will be returned, and will be shown in a graph.